Substitute in directories and use regex.
We use a regex to only substitute if the template is at the end of the
path. So any file that has a needed substitution in a parent directory
will have that handled as we iterate through directories. The regex will
keep us from handling this again at a file level.